+++ backend/src/routes/authRoutes.js
@@ -5,6 +5,7 @@ const {
     register,
     registerBusiness,
     login,
+    forgotPassword,
     createStaffAccount,
     createManagerAccount,
 } = require("../controllers/authController");
@@ -14,6 +15,7 @@ const validate = require("../middleware/validateMiddleware");
 const {
     registerSchema,
     loginSchema,
+    forgotPasswordSchema,
 } = require("../validators/authValidator");
 
 const protect = require("../middleware/authMiddleware");
@@ -30,5 +32,6 @@ router.post("/register-business", registerBusiness);
 router.post("/create-staff", protect, createStaffAccount);
 router.post("/create-manager", protect, createManagerAccount);
 router.post("/login", validate(loginSchema), login);
+router.post("/forgot-password", validate(forgotPasswordSchema), forgotPassword);
 
 module.exports = router;